+++ neo/addons/auxiliary/utility.py
@@ -16,25 +16,7 @@ import neo
 if TYPE_CHECKING:
     from googletrans import Translator
 
-_language_codes = neo.formatters.Table()
-_language_codes.init_columns("Lang Code", "Lang")
-for code, lang in [
-    ("en_US", "US English"),
-    ("hi", "Hindi"),
-    ("es", "Spanish"),
-    ("fr", "French"),
-    ("ja", "Japanese"),
-    ("ru", "Russian"),
-    ("en_GB", "UK English"),
-    ("de", "German"),
-    ("it", "Italian"),
-    ("ko", "Korean"),
-    ("pt-BR", "Portuguese"),
-    ("ar", "Arabic"),
-    ("tr", "Turkish"),
-]:
-    _language_codes.add_row(code, lang)
-LANGUAGE_CODES = _language_codes.display()
+
 TRANSLATION_DIRECTIVE = re.compile(
     r"((?P<src>[a-zA-Z\*\_]+|auto)->(?P<dest>[a-zA-Z]+))?"
 )